+++ modules/braille/braille-common/src/main/java/org/daisy/pipeline/braille/common/BrailleTranslator.java
@@ -1,7 +1,6 @@
 package org.daisy.pipeline.braille.common;
 
 import org.daisy.dotify.api.translator.BrailleTranslatorResult;
-import org.daisy.pipeline.braille.css.CSSStyledText;
 
 public interface BrailleTranslator extends Transform {