+++ modules/braille/braille-common/src/main/java/org/daisy/pipeline/braille/common/util.java
@@ -277,10 +277,6 @@ public abstract class util {
 				protected Iterator<T> _iterator() {
 					return iterable.iterator();
 				}
-				@Override
-				public String toString() {
-					return iterable.toString();
-				}
 			};
 		}