+++ modules/braille/dotify-utils/src/main/java/org/daisy/pipeline/braille/dotify/impl/BrailleTranslatorFactoryServiceImpl.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
 import java.util.regex.Matcher;
 
 import com.google.common.base.Optional;
@@ -32,12 +31,11 @@ import org.daisy.dotify.api.translator.TranslatorSpecification;
 
 import org.daisy.pipeline.braille.common.AbstractBrailleTranslator.util.DefaultLineBreaker;
 import org.daisy.pipeline.braille.common.BrailleTranslatorRegistry;
+import org.daisy.pipeline.braille.common.CSSStyledText;
 import org.daisy.pipeline.braille.common.Query;
 import static org.daisy.pipeline.braille.common.Query.util.mutableQuery;
 import static org.daisy.pipeline.braille.common.Query.util.query;
 import static org.daisy.pipeline.braille.common.Query.util.QUERY;
-import static org.daisy.pipeline.braille.common.util.Locales.parseLocale;
-import org.daisy.pipeline.braille.css.CSSStyledText;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -161,10 +159,7 @@ public class BrailleTranslatorFactoryServiceImpl implements BrailleTranslatorFac
 					// Because normally this will never end up in the resulting PEF, it is okay to replace
 					// it with "0". We choose a number because in some cases the translator may assume that
 					// the input is numerical.
-					styledText.set(i, new CSSStyledText("0",
-					                                    styledText.get(i).getStyle(),
-					                                    styledText.get(i).getLanguage(),
-					                                    styledText.get(i).getTextAttributes()));
+					styledText.set(i, new CSSStyledText("0", styledText.get(i).getStyle()));
 			return translate(styledText, from, to);
 		}
 		
@@ -189,18 +184,15 @@ public class BrailleTranslatorFactoryServiceImpl implements BrailleTranslatorFac
 	 * property is assumed to have been performed already.
 	 */
 	protected static Iterable<CSSStyledText> cssStyledTextFromTranslatable(Translatable specification) {
-		String locale = specification.getLocale();
 		return cssStyledTextFromTranslatable(
 			specification.getText(),
 			specification.getAttributes(),
-			locale != null ? parseLocale(locale) : null,
 			specification.isHyphenating(),
 			null);
 	}
 
 	private static Iterable<CSSStyledText> cssStyledTextFromTranslatable(String text,
 	                                                                     TextAttribute attributes,
-	                                                                     Locale language,
 	                                                                     boolean hyphenating,
 	                                                                     SimpleInlineStyle parentStyle) {
 		if (attributes != null && attributes.getWidth() != text.length())
@@ -216,21 +208,20 @@ public class BrailleTranslatorFactoryServiceImpl implements BrailleTranslatorFac
 				// FIXME: add hyphens declaration through braille-css model
 				style = new SimpleInlineStyle("hyphens: auto", style); }
 		if (attributes != null && attributes.hasChildren())
-			return cssStyledTextFromTranslatable(text, attributes.iterator(), language, false, style);
+			return cssStyledTextFromTranslatable(text, attributes.iterator(), false, style);
 		else
-			return Collections.singleton(new CSSStyledText(text, style, language));
+			return Collections.singleton(new CSSStyledText(text, style));
 	}
 
 	private static Iterable<CSSStyledText> cssStyledTextFromTranslatable(String text,
 	                                                                     Iterator<TextAttribute> attributes,
-	                                                                     Locale language,
 	                                                                     boolean hyphenating,
 	                                                                     SimpleInlineStyle parentStyle) {
 		if (attributes.hasNext()) {
 			TextAttribute a = attributes.next();
 			int w = a.getWidth();
-			return concat(cssStyledTextFromTranslatable(text.substring(0, w), a, language, hyphenating, parentStyle),
-			              cssStyledTextFromTranslatable(text.substring(w), attributes, language, hyphenating, parentStyle)); }
+			return concat(cssStyledTextFromTranslatable(text.substring(0, w), a, hyphenating, parentStyle),
+			              cssStyledTextFromTranslatable(text.substring(w), attributes, hyphenating, parentStyle)); }
 		else
 			return empty;
 	}
@@ -311,15 +302,12 @@ public class BrailleTranslatorFactoryServiceImpl implements BrailleTranslatorFac
 		List<CSSStyledText> styledText = new ArrayList<>();
 		for (Object t : precedingOrFollowingText) {
 			String text;
-			String locale;
 			boolean hyphenate; {
 				if (t instanceof PrecedingText) {
 					text = ((PrecedingText)t).resolve();
-					locale = ((PrecedingText)t).getLocale().orElse(null);
 					hyphenate = ((PrecedingText)t).shouldHyphenate(); }
 				else if (t instanceof FollowingText) {
 					text = ((FollowingText)t).peek();
-					locale = ((FollowingText)t).getLocale().orElse(null);
 					hyphenate = ((FollowingText)t).shouldHyphenate(); }
 				else
 					throw new RuntimeException();
@@ -332,7 +320,7 @@ public class BrailleTranslatorFactoryServiceImpl implements BrailleTranslatorFac
 				else
 					style = new SimpleInlineStyle("", parentStyle);
 			}
-			styledText.add(new CSSStyledText(text, style, locale != null ? parseLocale(locale) : null));
+			styledText.add(new CSSStyledText(text, style));
 		}
 		return styledText;
 	}