+++ modules/braille/libhyphen-utils/src/test/java/LibhyphenCoreTest.java
@@ -1,8 +1,6 @@
 import java.io.IOException;
 import java.io.File;
 import java.io.FileInputStream;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Properties;
 
 import javax.inject.Inject;
@@ -11,7 +9,6 @@ import static org.daisy.pipeline.braille.common.Hyphenator.FullHyphenator;
 import static org.daisy.pipeline.braille.common.Hyphenator.LineBreaker;
 import static org.daisy.pipeline.braille.common.Hyphenator.LineIterator;
 import static org.daisy.pipeline.braille.common.Query.util.query;
-import org.daisy.pipeline.braille.css.CSSStyledText;
 
 import org.daisy.pipeline.braille.libhyphen.LibhyphenHyphenator;
 
@@ -49,10 +46,9 @@ public class LibhyphenCoreTest extends AbstractTest {
 		                                   .get(query("(table:'standard.dic')"))
 		                                   .iterator().next()
 		                                   .asFullHyphenator();
-		assertEquals(text("foo\u00ADbar"), hyphenator.transform(styledText("foobar", "hyphens: auto")));
-		assertEquals(text("foo-\u200Bbar"), hyphenator.transform(styledText("foo-bar", "hyphens: auto")));
-		assertEquals(text("foo-\u200Bbar"), hyphenator.transform(styledText("foo-bar", "hyphens: none")));
-		assertEquals(text("foo\u00ADbar foob\u00ADar"), hyphenator.transform(styledText("foobar foob\u00ADar", "hyphens: auto")));
+		assertEquals("foo\u00ADbar", hyphenator.transform("foobar"));
+		assertEquals("foo-\u200Bbar", hyphenator.transform("foo-bar"));
+		assertEquals("foo\u00ADbar foob\u00ADar", hyphenator.transform("foobar foob\u00ADar"));
 	}
 	
 	@Test(expected=RuntimeException.class)
@@ -61,7 +57,7 @@ public class LibhyphenCoreTest extends AbstractTest {
 		                                   .get(query("(table:'non-standard.dic')"))
 		                                   .iterator().next()
 		                                   .asFullHyphenator();
-		hyphenator.transform(styledText("foobar", "hyphens: auto"));
+		hyphenator.transform("foobar");
 	}
 	
 	@Test
@@ -74,41 +70,19 @@ public class LibhyphenCoreTest extends AbstractTest {
 		             "oo\n" +
 		             "ba\n" +
 		             "r",
-		             fillLines(hyphenator.transform("foobar", null), 2, '-'));
+		             fillLines(hyphenator.transform("foobar"), 2, '-'));
 		assertEquals("fu-\n" +
 		             "bar",
-		             fillLines(hyphenator.transform("foobar", null), 3, '-'));
+		             fillLines(hyphenator.transform("foobar"), 3, '-'));
 		assertEquals("foo-\n" +
 		             "bar",
-		             fillLines(hyphenator.transform("foo-bar", null), 4, '-'));
+		             fillLines(hyphenator.transform("foo-bar"), 4, '-'));
 		assertEquals("foo-\n" +
 		             "bar",
-		             fillLines(hyphenator.transform("foo-bar", null), 5, '-'));
+		             fillLines(hyphenator.transform("foo-bar"), 5, '-'));
 		assertEquals("foo-\n" +
 		             "bar",
-		             fillLines(hyphenator.transform("foo-bar", null), 6, '-'));
-	}
-	
-	private Iterable<CSSStyledText> styledText(String... textAndStyle) {
-		List<CSSStyledText> styledText = new ArrayList<CSSStyledText>();
-		String text = null;
-		boolean textSet = false;
-		for (String s : textAndStyle) {
-			if (textSet)
-				styledText.add(new CSSStyledText(text, s));
-			else
-				text = s;
-			textSet = !textSet; }
-		if (textSet)
-			throw new RuntimeException();
-		return styledText;
-	}
-	
-	private Iterable<CSSStyledText> text(String... text) {
-		List<CSSStyledText> styledText = new ArrayList<CSSStyledText>();
-		for (String t : text)
-			styledText.add(new CSSStyledText(t, ""));
-		return styledText;
+		             fillLines(hyphenator.transform("foo-bar"), 6, '-'));
 	}
 	
 	private static String fillLines(LineIterator lines, int width, char hyphen) {