+++ modules/braille/liblouis-utils/src/test/java/MockHyphenator.java
@@ -1,5 +1,3 @@
-import java.util.Locale;
-
 import com.google.common.base.MoreObjects;
 import com.google.common.base.MoreObjects.ToStringHelper;
 
@@ -9,14 +7,13 @@ import org.daisy.pipeline.braille.common.AbstractTransformProvider;
 import org.daisy.pipeline.braille.common.HyphenatorProvider;
 import org.daisy.pipeline.braille.common.Query;
 import static org.daisy.pipeline.braille.common.Query.util.query;
-import org.daisy.pipeline.braille.css.CSSStyledText;
 
 import org.osgi.service.component.annotations.Component;
 
 public class MockHyphenator extends AbstractHyphenator {
 	
 	private static final LineBreaker lineBreaker = new DefaultLineBreaker() {
-		protected Break breakWord(String word, Locale language, int limit, boolean force) {
+		protected Break breakWord(String word, int limit, boolean force) {
 			if (limit >= 3 && word.equals("foobarz"))
 				return new Break("fubbarz", 3, true);
 			else if (limit >= word.length())
@@ -29,7 +26,10 @@ public class MockHyphenator extends AbstractHyphenator {
 	};
 	
 	private static final FullHyphenator fullHyphenator = new FullHyphenator() {
-		public Iterable<CSSStyledText> transform(Iterable<CSSStyledText> text) throws NonStandardHyphenationException {
+		public String transform(String text) throws NonStandardHyphenationException {
+			throw new NonStandardHyphenationException();
+		}
+		public String[] transform(String[] text) throws NonStandardHyphenationException {
 			throw new NonStandardHyphenationException();
 		}
 	};