+++ modules/braille/NEWS.md
@@ -7,6 +7,22 @@ Changes
 - Allow disabling hyphenation completely (also where there are soft hyphens)
 - Default handling of lists, including support for `start` and `value` attributes
 - Changed default suffix of counter styles to ". "
+- Deprecated the support for XSLT braille style sheets. It is now advised to apply any XSLT style
+  sheets during pre-processing.
+- Improved hyphenation of Norwegian text.
+- Various bugfixes
+
+Components
+----------
+- liblouis ([3.21.0](https://github.com/liblouis/liblouis/releases/tag/v3.21.0)),
+  liblouisutdml ([2.5.0](https://github.com/liblouis/liblouisutdml/releases/tag/v2.5.0)),
+  **liblouis-java** ([**5.0.2**](https://github.com/liblouis/liblouis-java/releases/tag/5.0.1))
+- **dotify** ([**1.0.5**](https://github.com/mtmse/dotify.library/releases/tag/1.0.3)
+- braille-css ([1.21.0](https://github.com/daisy/braille-css/releases/tag/1.21.0))
+- jsass ([5.10.4-p1](https://github.com/snaekobbi/jsass/releases/tag/5.10.4-p1))
+- libhyphen ([2.8.8](https://github.com/snaekobbi/libhyphen-nar/releases/tag/2.8.8)), jhyphen
+  ([1.0.2](https://github.com/daisy/jhyphen/releases/tag/1.0.2))
+- texhyphj ([1.2](https://github.com/joeha480/texhyphj/releases/tag/release-1.2))
 
 v1.14.7
 =======